third_party/WebKit/LayoutTests/inspector/sources/outline-javascript-2.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/sources-test.js"></script>

<script>

async function test()
{
    var test = InspectorTest.testJavascriptOutline;

    InspectorTest.markStep('testNamedFunctionVariableAssignment');
    await test('var a = function foo(...bar) { }');

    InspectorTest.markStep('testArrowFunction');
    await test('var a = x => x + 2');

    InspectorTest.markStep('testArrowFunctionWithMultipleArguments');
    await test('var a = (x, y) => x + y');

    InspectorTest.markStep('testInnerFunctions');
    await test('function foo(){ function bar() {} function baz() { }}');

    InspectorTest.markStep('testObjectProperties');
    await test('x = { run: function() { }, get count() { }, set count(value) { }}');

    InspectorTest.completeTest();
}

</script>

</head>

<body onload="runTest()">
<p>Verify javascript outline</p>
</body>
</html>
